+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 ### Added
+- Support `TableViewManager` as `manager` property of `Item` and `Section`
 - Support `UIScrollViewDelegate` from `TableViewManager` via `scrollDelegate`
 
 ### Changed